pyproject.toml
[project]
name = "obox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "fastmcp>=2.14.4",
    "pydantic-ai>=1.51.0",
]
